NN: Fix a problem when setting up master/slave that can occur because the directories are not there
git-svn-id: http://svn.zentyal.org/zentyal/trunk@16220 866bc4c1-54da-0310-a958-9d5a9a0e8d04

+++ client/usersandgroups/src/EBox/Ldap.pm
@@ -867,13 +867,14 @@ sub _loadLdap
     my $ldifFile = $self->ldifFile($dir, $slapd, $type);
 
     my $backupCommand = $self->_backupSystemDirectory();
+    my $mkCommand = $self->_mkLdapDirCmd($ldapDir);
     my $rmCommand = $self->_rmLdapDirCmd($ldapDir);
     my $slapaddCommand = $self->_slapaddCmd($ldifFile, $slapd, $type);
     my $chownConfCommand = $self->_chownConfDir($slapd);
     my $chownDataCommand = $self->_chownDataDir($slapd);
 
     $self->_execute(
-                cmds => [$backupCommand, $rmCommand,
+                cmds => [$backupCommand, $mkCommand, $rmCommand,
                          $slapaddCommand, $chownConfCommand, $chownDataCommand
                         ]);
 }
@@ -1007,12 +1008,23 @@ sub _slapaddCmd
     return  "/usr/sbin/slapadd -c $options -F " . confDir($slapd) .  " -b '$base' < $ldifFile";
 }
 
+sub _mkLdapDirCmd
+{
+    my ($self, $ldapDir)   = @_;
+
+    return "mkdir -p $ldapDir";
+}
+
 sub _rmLdapDirCmd
 {
     my ($self, $ldapDir)   = @_;
-    $ldapDir .= '/*' if defined $ldapDir ;
 
-    return "sh -c '/bin/rm -rf $ldapDir'";
+    if (defined($ldapDir)) {
+        $ldapDir .= '/*';
+        return "sh -c '/bin/rm -rf $ldapDir'";
+    } else {
+        return "true";
+    }
 }
 
 sub _backupSystemDirectory
